feat: complete data-driven framework for User, Pet, Store modules
- All test cases passing across all modules

- Allure reporting integrated (results ignored in git)

- Excel-based test data management

- Full CRUD coverage for all endpoints

- Type conversion handled in test layer

+++ src/test/java/api/test/StoreTest.java
@@ -14,7 +14,7 @@ public class StoreTest {
 	
 	Faker fake;
 	Store str;
-	int orderId;
+	long orderId;
 	
 	@BeforeClass
